-refactor getDarData into utils and extract repeated code from reviewResults, and accessReview
